requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
black==24.8.0
click==8.1.7
coverage==7.6.1
Django==4.2
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.7
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.4
graphql-relay==3.2.0
inflection==0.5.1
mypy-extensions==1.0.0
nose==1.3.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
promise==2.3
psycopg2-binary==2.9.9
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
